windows_train_model (single_GPU)/requirements.txt
adapters==0.2.2
einops==0.8.0
icecream==2.1.3
lightning==2.4.0
matplotlib==3.9.1.post1
numpy==2.0.1
pandas==2.2.2
Pillow==10.4.0
Requests==2.32.3
sacred==0.8.5
scikit_learn==1.5.1
scipy==1.14.0
skipthoughts==0.0.1
timm==1.0.8
tqdm==4.66.4
transformers==4.40.2
wandb